backend/requirements.txt
fastapi
uvicorn
pydantic[dotenv]
httpx
python-dotenv

